backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pymongo==4.12.1
pydantic==2.11.3
python-dotenv==1.1.0